Detailed information on CHELSA climate variables

Detailed information on CHELSA climate variables

Usage

CHELSA_Vars

Format

An object of class tbl_df (inherits from tbl, data.frame) with 46 rows and 6 columns.

Value

A tibble containing detailed information about various climate variables in CHELSA climate data. Each row in the tibble represents a different climate variable, with columns providing additional details such as the long name, unit, scale, offset, and an explanation of the variable

Examples

data(CHELSA_Vars)
print(CHELSA_Vars, n = Inf)
#> # A tibble: 46 × 6
#>    Variable Long_name                             unit  scale offset explanation
#>    <chr>    <chr>                                 <chr> <dbl>  <dbl> <chr>      
#>  1 bio1     mean annual air temperature           deg C   0.1  -273. "mean annu…
#>  2 bio2     mean diurnal air temperature range    deg C   0.1     0  "mean diur…
#>  3 bio3     isothermality                         deg C   0.1     0  "ratio of …
#>  4 bio4     temperature seasonality               deg …   0.1     0  "standard …
#>  5 bio5     mean daily maximum air temperature o… deg C   0.1  -273. "The highe…
#>  6 bio6     mean daily minimum air temperature o… deg C   0.1  -273. "The lowes…
#>  7 bio7     annual range of air temperature       deg C   0.1     0  "Differenc…
#>  8 bio8     mean daily mean air temperatures of … deg C   0.1  -273. "The wette…
#>  9 bio9     mean daily mean air temperatures of … deg C   0.1  -273. "The dries…
#> 10 bio10    mean daily mean air temperatures of … deg C   0.1  -273. "The warme…
#> 11 bio11    mean daily mean air temperatures of … deg C   0.1  -273. "The colde…
#> 12 bio12    annual precipitation amount           kg m…   0.1     0  "Accumulat…
#> 13 bio13    precipitation amount of the wettest … kg m…   0.1     0  "The preci…
#> 14 bio14    precipitation amount of the driest m… kg m…   0.1     0  "The preci…
#> 15 bio15    precipitation seasonality             kg m…   0.1     0  "The Coeff…
#> 16 bio16    mean monthly precipitation amount of… kg m…   0.1     0  "The wette…
#> 17 bio17    mean monthly precipitation amount of… kg m…   0.1     0  "The dries…
#> 18 bio18    mean monthly precipitation amount of… kg m…   0.1     0  "The warme…
#> 19 bio19    mean monthly precipitation amount of… kg m…   0.1     0  "The colde…
#> 20 gdgfgd0  First growing degree day above 0 deg… juli…   1       0  "First day…
#> 21 gdgfgd5  First growing degree day above 5 deg… juli…   1       0  "First day…
#> 22 gdgfgd10 First growing degree day above 10 de… juli…   1       0  "First day…
#> 23 gddlgd0  Last growing degree day above 0 deg C juli…   1       0  "Last day …
#> 24 gddlgd5  Last growing degree day above 5 deg C juli…   1       0  "Last day …
#> 25 gddlgd10 Last growing degree day above 10 deg… juli…   1       0  "Last day …
#> 26 gdd0     Growing degree days heat sum above 0… deg C   0.1     0  "heat sum …
#> 27 gdd5     Growing degree days heat sum above 5… deg C   0.1     0  "heat sum …
#> 28 gdd10    Growing degree days heat sum above 1… deg C   0.1     0  "heat sum …
#> 29 ngd0     Number of growing degree days         # da…   1       0  "Number of…
#> 30 ngd5     Number of growing degree days         # da…   1       0  "Number of…
#> 31 ngd10    Number of growing degree days         # da…   1       0  "Number of…
#> 32 gsl      growing season length TREELIM         # da…   1       0  "Length of…
#> 33 gst      Mean temperature of the growing seas… deg C   0.1  -273. "Mean temp…
#> 34 lgd      last day of the growing season TREEL… juli…   1       0  "Last day …
#> 35 fgd      first day of the growing season TREE… juli…   1       0  "first day…
#> 36 gsp      Accumulated precipitation amount on … kg m…   0.1     0  "precipita…
#> 37 fcf      Frost change frequency                count   1       0  "Number of…
#> 38 scd      Snow cover days                       count   1       0  "Number of…
#> 39 swe      Snow water equivalent                 kg m…   0.1     0  "Amount of…
#> 40 kg0      Koeppen-Geiger climate classification cate…   1       0  "Koeppen G…
#> 41 kg1      Koeppen-Geiger climate classification cate…   1       0  "Koeppen G…
#> 42 kg2      Koeppen-Geiger climate classification cate…   1       0  "Koeppen G…
#> 43 kg3      Koeppen-Geiger climate classification cate…   1       0  "Wissmann …
#> 44 kg4      Koeppen-Geiger climate classification cate…   1       0  "Thornthwa…
#> 45 kg5      Koeppen-Geiger climate classification cate…   1       0  "Troll-Pfa…
#> 46 npp      Net primary productivity              g C …   0.1     0  "Calculate…