Detailed information on CHELSA climate variables

Detailed information on CHELSA climate variables

Usage

CHELSA_variables

Format

An object of class tbl_df (inherits from tbl, data.frame) with 46 rows and 6 columns.

Value

A tibble containing detailed information about various climate variables in CHELSA climate data. Each row in the tibble represents a different climate variable, with columns providing additional details such as the long name, unit, scale, offset, and an explanation of the variable

Examples

ecokit::load_packages(dplyr, gt)

data(CHELSA_variables)
gt::gt(CHELSA_variables)

  

Variable
      	Long_name
      	unit
      	scale
      	offset
      	explanation
    

bio1
	mean annual air temperature
	deg C
	0.1
	-273.15
	mean annual daily mean air temperatures averaged over 1 year

bio2
	mean diurnal air temperature range
	deg C
	0.1
	0.00
	mean diurnal range of temperatures averaged over 1 year

bio3
	isothermality
	deg C
	0.1
	0.00
	ratio of diurnal variation to annual variation in temperatures

bio4
	temperature seasonality
	deg C/100
	0.1
	0.00
	standard deviation of the monthly mean temperatures

bio5
	mean daily maximum air temperature of the warmest month
	deg C
	0.1
	-273.15
	The highest temperature of any monthly daily mean maximum temperature

bio6
	mean daily minimum air temperature of the coldest month
	deg C
	0.1
	-273.15
	The lowest temperature of any monthly daily mean minimum temperature

bio7
	annual range of air temperature
	deg C
	0.1
	0.00
	Difference between the Maximum Temperature of Warmest month and the Minimum Temperature of Coldest month (bio5-bio6)

bio8
	mean daily mean air temperatures of the wettest quarter
	deg C
	0.1
	-273.15
	The wettest quarter of the year is determined (to the nearest month)

bio9
	mean daily mean air temperatures of the driest quarter
	deg C
	0.1
	-273.15
	The driest quarter of the year is determined (to the nearest month)

bio10
	mean daily mean air temperatures of the warmest quarter
	deg C
	0.1
	-273.15
	The warmest quarter of the year is determined (to the nearest month)

bio11
	mean daily mean air temperatures of the coldest quarter
	deg C
	0.1
	-273.15
	The coldest quarter of the year is determined (to the nearest month)

bio12
	annual precipitation amount
	kg m-2
	0.1
	0.00
	Accumulated precipitation amount over 1 year

bio13
	precipitation amount of the wettest month
	kg m-2
	0.1
	0.00
	The precipitation of the wettest month.

bio14
	precipitation amount of the driest month
	kg m-2
	0.1
	0.00
	The precipitation of the driest month.

bio15
	precipitation seasonality
	kg m-2
	0.1
	0.00
	The Coefficient of Variation = 100*standard deviation of the monthly precipitation / mean 

bio16
	mean monthly precipitation amount of the wettest quarter
	kg m-2
	0.1
	0.00
	The wettest quarter of the year is determined (to the nearest month)

bio17
	mean monthly precipitation amount of the driest quarter
	kg m-2
	0.1
	0.00
	The driest quarter of the year is determined (to the nearest month)

bio18
	mean monthly precipitation amount of the warmest quarter
	kg m-2
	0.1
	0.00
	The warmest quarter of the year is determined (to the nearest month)

bio19
	mean monthly precipitation amount of the coldest quarter
	kg m-2
	0.1
	0.00
	The coldest quarter of the year is determined (to the nearest month)

gdgfgd0
	First growing degree day above 0 deg C
	julian day
	1.0
	0.00
	First day of the year above 0 deg C

gdgfgd5
	First growing degree day above 5 deg C
	julian day
	1.0
	0.00
	First day of the year above 5 deg C

gdgfgd10
	First growing degree day above 10 deg C
	julian day
	1.0
	0.00
	First day of the year above 10 deg C

gddlgd0
	Last growing degree day above 0 deg C
	julian day
	1.0
	0.00
	Last day of the year above 0 deg C

gddlgd5
	Last growing degree day above 5 deg C
	julian day
	1.0
	0.00
	Last day of the year above 5 deg C

gddlgd10
	Last growing degree day above 10 deg C
	julian day
	1.0
	0.00
	Last day of the year above 10 deg C

gdd0
	Growing degree days heat sum above 0 deg C
	deg C
	0.1
	0.00
	heat sum of all days above the 0 deg C temperature accumulated over 1 year.

gdd5
	Growing degree days heat sum above 5 deg C
	deg C
	0.1
	0.00
	heat sum of all days above the 5 deg C temperature accumulated over 1 year.

gdd10
	Growing degree days heat sum above 10 deg C
	deg C
	0.1
	0.00
	heat sum of all days above the 10 deg C temperature accumulated over 1 year.

ngd0
	Number of growing degree days
	# days
	1.0
	0.00
	Number of days at which tas > 0 deg C

ngd5
	Number of growing degree days
	# days
	1.0
	0.00
	Number of days at which tas > 5 deg C

ngd10
	Number of growing degree days
	# days
	1.0
	0.00
	Number of days at which tas > 10 deg C

gsl
	growing season length TREELIM
	# days
	1.0
	0.00
	Length of the growing season

gst
	Mean temperature of the growing season TREELIM
	deg C
	0.1
	-273.15
	Mean temperature of all growing season days based on TREELIM

lgd
	last day of the growing season TREELIM
	julian day
	1.0
	0.00
	Last day of the growing season according to TREELIM

fgd
	first day of the growing season TREELIM
	julian day
	1.0
	0.00
	first day of the growing season according to TREELIM

gsp
	Accumulated precipitation amount on growing season days TREELIM
	kg m-2
	0.1
	0.00
	precipitation sum accumulated on all days during the growing season based on TREELIM

fcf
	Frost change frequency
	count
	1.0
	0.00
	Number of events in which tmin or tmax go above, or below 0 deg C

scd
	Snow cover days
	count
	1.0
	0.00
	Number of days with snowcover calculated using the snowpack model implementation in from TREELIM

swe
	Snow water equivalent
	kg m-2
	0.1
	0.00
	Amount of luquid water if snow is melted

kg0
	Koeppen-Geiger climate classification
	category
	1.0
	0.00
	Koeppen Geiger - Koeppen&Geiger (1936)

kg1
	Koeppen-Geiger climate classification
	category
	1.0
	0.00
	Koeppen Geiger without As/Aw differentiation - Koeppen&Geiger (1936)

kg2
	Koeppen-Geiger climate classification
	category
	1.0
	0.00
	Koeppen Geiger after Peel et al. 2007

kg3
	Koeppen-Geiger climate classification
	category
	1.0
	0.00
	Wissmann 1939

kg4
	Koeppen-Geiger climate classification
	category
	1.0
	0.00
	Thornthwaite 1931

kg5
	Koeppen-Geiger climate classification
	category
	1.0
	0.00
	Troll-Pfaffen - Troll&Paffen (1964)

npp
	Net primary productivity
	g C m-2 yr-1
	0.1
	0.00
	Calculated based on the `Miami model`, Lieth, H., 1972.