Process Corine land cover (CLC) data

Processes Corine Land Cover (CLC) data for environmental modeling purposes. The function calculates the percentage coverage of CLC per grid cell. It estimate the percent coverage at the 3 CLC levels, and cross-walk for EUNIS_19 and SynHab habitat types. Similarly, the function estimates the most common class per grid cell (3 CLC levels, EUNIS, and SynHab habitat types) and prepares reference grid for models. The function optionally plots % coverage maps.

Usage

CLC_Process(EnvFile = ".env", FromHPC = TRUE, MinLandPerc = 15, PlotCLC = TRUE)

Arguments

EnvFile

String specifying the path to the environment file containing necessary paths and configurations.

FromHPC

Logical indicating whether the processing is being done on a High-Performance Computing (HPC) environment. Defaults to TRUE.

MinLandPerc

A numeric value indicating the minimum percentage of land per grid cell to be used in the reference grid cell. Defaults to 15.

PlotCLC

Logical indicating whether to plot the percentage coverage of different levels of CLC and custom habitat types. Defaults to TRUE.

Value

The function does not return a value but produces side effects such as saving processed data and plots to specified directories.

Details

Data source

Citation

The function reads the following environment variable:

DP_R_Grid (if FromHPC = TRUE) or DP_R_Grid_Local (if FromHPC = FALSE): Path for saving the processed CLC data.

DP_R_Grid_Ref (if FromHPC = TRUE) or DP_R_Grid_Ref_Local (if FromHPC = FALSE). The function reads the content of Grid_10_sf.RData and Grid_10_Raster.RData files from this path.

DP_R_EUBound_sf (if FromHPC = TRUE) or DP_R_EUBound_sf_Local (if FromHPC = FALSE): path for the RData file containing the country boundaries (sf object).

DP_R_CLC (if FromHPC = TRUE) or DP_R_CLC_Local (if FromHPC = FALSE): directory where the outputs of CLC data processing are processed.

DP_R_CLC_CW (if FromHPC = TRUE) or DP_R_CLC_CW_Local (if FromHPC = FALSE): path for the crossWalk.txt file containing custom cross-walk between CLC values and their corresponding values for three levels of CLC and EUNIS_19 & SynHab habitat types.

DP_R_CLC_tif (if FromHPC = TRUE) or DP_R_CLC_tif_Local (if FromHPC = FALSE): path for the input CLC .tif file.

Author

Ahmed El-Gabbas