Check the Integrity of an qs2 File

This function checks if the specified file is an qs2 file and attempts to read it using qs2::qs_read. It returns TRUE if the file is an qs2 file and contains a non-null object, otherwise FALSE.

Usage

CheckQs(File, warning = TRUE, nthreads = 5)

Arguments

File

A character string specifying the path to the file to be checked. This can not be empty. The function will attempt to determine the file type based on the file extension. If the file extension is not qs2, the function will return FALSE.

warning

logical. If TRUE (default), warnings will be printed if the file does not exist.

nthreads

integer. The number of threads to use when reading the qs2 file. Default is 5.

Value

A logical value: TRUE if the file is an qs2 file and contains a non-null object, FALSE otherwise.

Author

Ahmed El-Gabbas