Check the Integrity of a ZIP File

This function checks if a given ZIP file exists and whether it is intact without any errors. The function uses the system's unzip command to test the file's integrity.

Usage

CheckZip(File)

Arguments

File

Character. The path to the ZIP file that needs to be checked.

Value

Logical. Returns TRUE if the file exists and passes the integrity check, and FALSE otherwise.

Author

Ahmed El-Gabbas