Plot model convergence of a selected model

This function generates and saves plots for model convergence diagnostics, including rho, alpha, omega, and beta parameters. It supports parallel processing for faster execution and can work with models fitted on High-Performance Computing HPC environments.

Usage

Convergence_Plot(
  Path_Coda = NULL,
  Path_Model = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  Title = " ",
  NOmega = 1000,
  NCores = 8,
  NRC = c(2, 2),
  Beta_NRC = c(3, 3),
  SavePlotData = TRUE,
  PagePerFile = 20,
  Cols = NULL,
  MarginType = "histogram"
)

Arguments

Path_Coda

String. Path to the coda object containing MCMC samples.

Path_Model

String. Path to the fitted Hmsc model object.

EnvFile

String. Path to the environment file containing necessary environment variables. Default: ".env".

FromHPC

Logical. Indicates whether the function is being run on an HPC environment, affecting file path handling. Default: TRUE.

Title

String. Title for the rho and alpha plots. Default: " ".

NOmega

Integer. Number of species interactions to sample for omega parameter convergence diagnostics. Default: 1000.

NCores

Integer. Number of cores to use for parallel processing. Default: 8.

NRC

Numeric Vector. Specifies the number of rows and columns for arranging multiple plots on a page. Default: c(2, 2).

Beta_NRC

Numeric Vector. Specifies the number of rows and columns for arranging panels for the beta parameter. Default: c(3, 3).

SavePlotData

Logical. Indicates whether to save the plot data as RData files. Default: TRUE.

PagePerFile

Integer. Indicates the number of pages per single pdf page for the Omega parameter. Default: 20.

Cols

Character vector for chain colours (optional). Default: NULL.

MarginType

A string specifying the type of marginal plot to add to the main plot. Must be either "histogram" or "density". Default: "histogram".

Value

The function does not return a value but generates and saves plots to disk.

Author

Ahmed El-Gabbas