Request and manage GBIF data for vascular plant orders

This function requests GBIF data for each vascular plant order, processes it in parallel, and manages the data download and storage. If data is already available, it loads the data instead of making new requests.

Usage

Efforts_Request(
  NCores = 3,
  Path_Requests,
  Path_Efforts,
  StartYear = 1981,
  Boundaries = c(-30, 50, 25, 75)
)

Arguments

NCores

Integer. The number of cores to use for parallel processing (between 1 and 3). Must be a positive integer. Defaults to 3.

Path_Requests

Character. The directory path to save individual request files.

Path_Efforts

Character. The directory path to save the final compiled data.

StartYear

Numeric. The starting year for the occurrence data. Only records from this year onward will be requested from GBIF. Default is 1981, which matches the year ranges of CHELSA current climate data.

Boundaries

Numeric vector of length 4. Specifies geographical boundaries for the requested GBIF data in the order: Left, Right, Bottom, Top. Defaults to c(-30, 50, 25, 75).

Value

The function returns the GBIF data requests processed and stored in the specified directories.

Note

This function is not intended to be used directly by the user or in the IAS-pDT, but only used inside the Efforts_Process function.

Author

Ahmed El-Gabbas