Process and summarize GBIF data for vascular plants

This function processes GBIF data for vascular plants by extracting, summarizing, and saving the data. It also creates summary maps and saves them in both RData and TIFF formats.

Usage

Efforts_Summarize(
  NCores,
  Path_Efforts,
  Path_Efforts_Interim,
  Path_Efforts_Data,
  Path_Grid,
  IAS_List,
  Efforts_AllRequests,
  ChunkSize = 100000L,
  DeleteChunks = TRUE
)

Arguments

NCores

Integer. Number of CPU cores to use for parallel processing.

Path_Efforts

Character. Path where the final processed data will be saved.

Path_Efforts_Interim

Character. The directory path to save interim data.

Path_Efforts_Data

Character. The directory path to save detailed effort data as RData.

Path_Grid

Character. The directory path to load the grid data.

IAS_List

A list of invasive alien species keys.

Efforts_AllRequests

A data frame containing the details of the GBIF download, including paths to CSV files, zip files, order, class, and total records.

ChunkSize

Integer. The number of rows per chunk file. Default: 100,000. See Efforts_Split for more details.

DeleteChunks

logical, indicating whether to remove file chunks after processing the data. Defaults to TRUE.

Value

The function returns and saves the GBIF data summary.

Note

This function is not intended to be used directly by the user or in the IAS-pDT, but only used inside the Efforts_Process function.

Author

Ahmed El-Gabbas