Summarize maps for efforts data

This function processes spatial data (as an sf object), summarizes it based on the number of observations or distinct species, and generates a raster layer.

Usage

Efforts_SummarizeMaps(Data, NSp, Name, ClassOrder, Grid_SF, Grid_R)

Arguments

Data

An sf object containing spatial data, with a column named CellCode.

NSp

A logical value indicating whether to generate distinct species counts (TRUE) or total observation counts (FALSE).

Name

A character string that represents the name of the count field and the prefix for the final raster layer's name.

ClassOrder

A character string specifying the class and order combination (separated by an underscore) represented in the Data.

Grid_SF, Grid_R

Reference grid in the form of simple feature and raster.

Value

A processed terra raster object representing the summarized data.

Note

This function is not intended to be used directly by the user or in the IAS-pDT, but only used inside the Efforts_Process and Efforts_Summarize functions.

Author

Ahmed El-Gabbas