Read GBIF chunk data

This function reads and processes a chunk of GBIF (Global Biodiversity Information Facility) data. It filters the data based on several criteria including spatial uncertainty, collection year, coordinate precision, and taxonomic rank and returns only selected columns.

Usage

GBIF_ReadChunk(
  ChunkFile,
  EnvFile = ".env",
  FromHPC = TRUE,
  MaxUncert = 10,
  StartYear = 1981,
  SaveRData = TRUE,
  ReturnData = FALSE,
  Overwrite = FALSE
)

Arguments

ChunkFile

A string specifying the path of the chunk file to be processed.

EnvFile

A string specifying the path of the environment file. Default is ".env".

FromHPC

A logical value indicating whether the environment variables should be read from an HPC (High-Performance Computing) environment. Default is TRUE.

MaxUncert

A numeric value specifying the maximum accepted spatial uncertainty in kilometers. Default is 10 km.

StartYear

An integer specifying the earliest collection year to be included. Default is 1981.

SaveRData

Logical; indicating whether to save the cleaned data for the current chunk as *.RData file.

ReturnData

Logical; indicating whether to return the cleaned data for the current chunk. Defaults to FALSE.

Overwrite

Logical; indicating whether to process the current chunk file if it has already processed and saved as *.RData file. This helps to continue working on previously processed chunks if the previous try failed, e.g. due to memory issue.

Value

The output of the function is a tibble (sf) object containing the processed and filtered GBIF data for the input chunk file. Whether the tibble is exported or saved as *.RData file depends on the values of the SaveRData, ReturnData, and Overwrite parameters.

If SaveRData = TRUE (default), the processed data will be saved as RData file with the same base name of the chunk file and at the same directory. By default, the function does not return any value, unless ReturnData is set to TRUE. SaveRData and ReturnData can not both set to FALSE. The function can optionally skip processing the current chunk file if the *.RData file for this chunk already exist and Overwrite is set as FALSE (default). If Overwrite is set to TRUE, the data for the current chunk will be re-processed irrespective of the existence of the RData file. This helps to continue working on previously processed chunks if a previous run of GBIF_Process failed, e.g. due to memory issue.

Note

This function is not intended to be used directly by the user or in the IAS-pDT, but only used inside the GBIF_Process function.

Author

Ahmed El-Gabbas