Extract species data as sf and Raster and plot species distribution

Extracts and processes species data from processed GBIF data, generating grids, rasters, and visual maps.

Usage

GBIF_SpData(
  Species = NULL,
  FromHPC = TRUE,
  EnvFile = ".env",
  Verbose = TRUE,
  LastUpdate = NULL
)

Arguments

Species

A character string specifying the species name.

FromHPC

A logical value indicating whether the environment variables should be read from an HPC (High-Performance Computing) environment. Default is TRUE.

EnvFile

A string specifying the path of the environment file. Default is ".env".

Verbose

A logical value indicating whether to print progress messages. Default is TRUE.

LastUpdate

A character string specifying the last update date to be displayed on the plots.

Note

This function is not intended to be used directly by the user or in the IAS-pDT, but only used inside the GBIF_Process function.

Author

Ahmed El-Gabbas