Plot spatial variation in site loadings of HMSC Models

Generate and save spatial variation in site loadings of HMSC Models' latent factors as a JPEG file.

Usage

Mod_Plot_LF(
  Path_Model = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  Plot_Width = 20,
  Plot_Height = 21
)

Arguments

Path_Model

Path to the model file.

EnvFile

Path to the environment variables file. Default is ".env".

FromHPC

Logical flag indicating whether to load environment variables from HPC. Default is TRUE.

Plot_Width

Numeric value specifying the width of the output plot in cm. Default is 20.

Plot_Height

Numeric value specifying the height of the output plot in cm. Default is 21.

Value

Invisibly returns NULL after saving the plot.

Author

Ahmed El-Gabbas