Plot species and level of invasion predictions as JPEG files using ggplot2

Generate predictions for species and habitat models and saves the output as JPEG files.

Usage

Mod_Predict_Plot(ModelDir, EnvFile = ".env", FromHPC = TRUE, NCores = 8)

Arguments

ModelDir

Path to the model directory containing predictions.

EnvFile

Path to the environment file (.env) for setting paths.

FromHPC

Boolean indicating whether the environment is an HPC system.

NCores

Integer specifying the number of cores to use for parallel processing. Defaults to 8.

Value

Saves prediction plots as JPEG files in the specified output directory.

Author

Ahmed El-Gabbas