Creates heatmaps of parameter estimates or posterior support values for species' residual association (Omega parameters).

This function generates heatmaps to visualize the parameter estimates or posterior support values of species' residual associations (Omega parameters). It is designed to work with model output files and produces two types of visualizations: one indicating the sign (positive or negative) of the associations and another showing the mean values of these associations.

Usage

PlotOmegaGG(Path_Model, supportLevel = 0.95, PlotWidth = 26, PlotHeight = 22.5)

Arguments

Path_Model

String. Path to the fitted Hmsc model object.

supportLevel

Numeric. The threshold for posterior support values used to determine which associations are strong enough to be plotted. Only associations with posterior support exceeding this threshold (or falling below 1 - threshold for negative associations) will be visualized. Defaults to 0.95.

PlotWidth, PlotHeight

Integer. Specifies the width and height of the generated plot in centimeters. Defaults to 26 x 22.5.

Value

Generates two JPEG files containing the heatmaps of Omega parameter: signs and mean values. These files are saved in a directory named 'Model_Postprocessing' within the parent directory of the provided model file path. The function itself returns NULL invisibly.

Author

Ahmed El-Gabbas