Predict habitat suitability of Hmsc model across different climate options

This function generates prediction maps of Hmsc models for current and future climate options. It also predicts an ensemble predictions for different climate models. For each species and overall species richness, the function exports three maps: mean, standard deviation (sd), and coefficient of variation (cov). For future predictions, the function generates maps for prediction anomaly (future - current). The function prepares data to be uploaded to the OPeNDAP data server for the use of the IAS-pDT Shiny App.

Usage

Predict_Maps(
  Path_Model = NULL,
  Hab_Abb = NULL,
  EnvFile = ".env",
  FromHPC = TRUE,
  NCores = 8L,
  Pred_Clamp = TRUE,
  Fix_Efforts = "q90",
  Fix_Rivers = "q90",
  Pred_NewSites = TRUE,
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_NCores = NCores,
  LF_Check = FALSE,
  LF_Temp_Cleanup = TRUE,
  LF_Only = FALSE,
  LF_Commands_Only = FALSE,
  Temp_Dir = "TEMP_Pred",
  Temp_Cleanup = TRUE,
  CC_Models = c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0",
    "UKESM1-0-LL"),
  CC_Scenario = c("ssp126", "ssp370", "ssp585")
)

Arguments

Path_Model

Character. Path to fitted Hmsc model object.

Hab_Abb

Character. Habitat abbreviation indicating the specific SynHab habitat type for which data will be prepared. Valid values are 0, 1, 2, 3, 4a, 4b, 10, 12a, 12b. For more details, see Pysek et al..

EnvFile

Character. Path to the environment file containing paths to data sources. Defaults to .env.

FromHPC

Logical. Whether the processing is being done on an High-Performance Computing (HPC) environment, to adjust file paths accordingly. Default: TRUE.

NCores

Integer. Number of CPU cores to use for parallel processing. Default: 8.

Pred_Clamp

Logical indicating whether to clamp the sampling efforts at a single value. If TRUE (default), the Fix_Efforts argument must be provided.

Fix_Efforts

Numeric or character. If Pred_Clamp = TRUE, the sampling efforts predictor with values U+02264 Fix_Efforts is fixed at Fix_Efforts during predictions. If numeric, the value is directly used (log10 scale). If character, it can be one of median, mean, max, or q90 (90% Quantile). Using max can reflect extreme values caused by rare, highly sampled locations (e.g., urban centers or popular natural reserves). While using 90% quantile avoid such extreme grid cells while still capturing areas with high sampling effort. This argument is mandatory when Pred_Clamp is set to TRUE.

Fix_Rivers

Numeric or character. Similar to Fix_Efforts, but for fixing the length of rivers. If numeric, the value is directly used (log10 scale). If character, it can be one of median, mean, max, q90 (90% quantile). It can be also NULL for not fixing the river length predictor. Defaults to q90.

Pred_NewSites

Logical. Whether to predict habitat suitability at new sites. Default: TRUE. Note: This parameter is temporary and will be removed in future updates.

UseTF

Logical. Whether to use TensorFlow for calculations. Defaults to TRUE.

TF_Environ

Character. Path to the Python environment. This argument is required if UseTF is TRUE.

TF_use_single

Logical. Whether to use single precision for the TensorFlow calculations. Defaults to FALSE.

LF_NCores

Integer. Number of cores to use for parallel processing of latent factor prediction. Defaults to 8L.

LF_Check

Logical. If TRUE, the function checks if the output files are already created and valid. If FALSE, the function will only check if the files exist without checking their integrity. Default is FALSE.

LF_Temp_Cleanup

Logical. Whether to delete temporary files in the Temp_Dir directory after finishing the LF predictions.

LF_Only

Logical. Whether to predict only the latent factor. This is useful for distributing processing load between GPU and CPU. When LF_Only = TRUE, latent factor prediction needs to be computed separately on GPU. When computations are finished on GPU, the function can later be rerun with LF_Only = FALSE (default) to predict habitat suitability using the already-computed latent factor predictions.

LF_Commands_Only

Logical. If TRUE, returns the command to run the Python script. Default is FALSE.

Temp_Dir

Character. Path for temporary storage of intermediate files.

Temp_Cleanup

Logical. Whether to clean up temporary files. Defaults to TRUE.

CC_Models

Character vector. Climate models for future predictions. Available options are c("GFDL-ESM4", "IPSL-CM6A-LR", "MPI-ESM1-2-HR", "MRI-ESM2-0", "UKESM1-0-LL") (default).

CC_Scenario

Character vector. Climate scenarios for future predictions. Available options are: c("ssp126", "ssp370", "ssp585") (default).

Value

A tibble containing the prediction summary and file paths for output *.tif files.

Author

Ahmed El-Gabbas