Calculate railway intensity based on OpenStreetMap data

This function downloads, processes, and analyzes railway data extracted from OpenRailwayMap available from OpenStreetMap Data Extracts. It supports parallel processing for faster execution and can calculate the total length of railways and distance to the nearest railway for each grid cell in Europe.

Usage

Railway_Intensity(
  FromHPC = TRUE,
  EnvFile = ".env",
  NCores = 6,
  DeleteProcessed = TRUE
)

Arguments

FromHPC

Logical indicating whether the work is being done from HPC, to adjust file paths accordingly. Default: TRUE.

EnvFile

Character. The path to the environment file containing variables required by the function. Default is ".env".

NCores

Numeric. Number of CPU cores to use for parallel processing. Default is 6.

DeleteProcessed

Logical indicating whether to delete the raw downloaded railways files after processing them. This helps to free large unnecessary file space (> 55 GB). Defaults to TRUE.

Value

NULL. Outputs processed files to the directories specified in the environment file.

Author

Ahmed El-Gabbas