Rename active geometry column of an sf object.

This function renames the active geometry column of a simple feature (sf) object to a new name provided by the user.

Usage

Rename_geometry(g = NULL, name = NULL)

Arguments

g

sf object; the simple feature object whose geometry column name is to be changed.

name

character; the new name for the geometry column.

Value

The modified sf object with the renamed geometry column.

References

Click here