Plot species response curves

Generates and saves response curve plots for species based on environmental variables and other factors. Plots show predicted habitat suitability across different values of each variable.

Usage

RespCurv_PlotSp(
  ModelDir = NULL,
  NCores = 20,
  EnvFile = ".env",
  FromHPC = TRUE,
  ReturnData = FALSE
)

Arguments

ModelDir

String. Path to the root directory of the fitted models without the trailing slash. The function reads data from RespCurv_DT subdirectory created by RespCurv_PrepData.

NCores

Integer. Number of cores to use for parallel processing. Defaults to 20.

EnvFile

String. Path to the environment variables file. Defaults to ".env".

FromHPC

Logical. Indicates whether the function is being run on an HPC environment, affecting file path handling. Default: TRUE.

ReturnData

Logical. Indicates whether the output data be returned as an R object.