Plot all species response curves

Generates and saves response curve plots for all species together in a single plot.

Usage

RespCurv_PlotSpAll(
  ModelDir = NULL,
  NCores = 8L,
  ReturnData = FALSE,
  PlotAlpha = 0.3
)

Arguments

ModelDir

String. Path to the root directory of the fitted models without the trailing slash. The function reads data from RespCurv_DT subdirectory created by RespCurv_PrepData.

NCores

Integer. Number of cores to use for parallel processing. Defaults to 20.

ReturnData

Logical. Indicates whether the output data be returned as an R object.

PlotAlpha

Numeric. The alpha value (transparency) for the response curve lines. Defaults to 0.3.

Author

Ahmed El-Gabbas