Prepare and process response curve data for Hmsc models

This function prepares and processes data for generating response curves for Hmsc models. It supports parallel processing and can return the processed data.

Usage

RespCurv_PrepData(
  Path_Model = NULL,
  N_Grid = 50,
  NCores = 8,
  ReturnData = FALSE,
  Probabilities = c(0.025, 0.5, 0.975),
  UseTF = TRUE,
  TF_Environ = NULL,
  TF_use_single = FALSE,
  LF_NCores = NCores,
  LF_Check = FALSE,
  LF_Temp_Cleanup = TRUE,
  LF_Commands_Only = FALSE,
  Temp_Dir = "TEMP_Pred",
  Temp_Cleanup = TRUE,
  Verbose = TRUE
)

Arguments

Path_Model

Character. Path file containing the model.

N_Grid

Integer specifying the number of points along the gradient for continuous focal variables. Defaults to 50. See Hmsc::constructGradient for more details.

NCores

Integer specifying the number of cores to use for parallel processing. Defaults to 8.

ReturnData

Logical. If TRUE, returns processed response curve data as an R object. Default is FALSE.

Probabilities

quantiles to be calculated. Defaults to c(0.025, 0.5, 0.975). See stats::quantile for more details.

UseTF

Logical indicating whether to use TensorFlow for calculations. Defaults to TRUE.

TF_Environ

Character string specifying the path to the Python environment. Defaults to NULL. This argument is required if UseTF is TRUE.

TF_use_single

Logical indicating whether to use single precision for the TF calculations. Defaults to FALSE.

LF_NCores

Integer specifying the number of cores to use for parallel processing. Defaults to 8.

LF_Check

Logical. If TRUE, the function checks if the output files are already created and valid. If FALSE, the function will only check if the files exist without checking their integrity. Default is FALSE.

LF_Temp_Cleanup

Logical indicating whether to delete temporary files in the Temp_Dir after finishing the LF predictions.

LF_Commands_Only

logical. If TRUE, returns the command to run the Python script. Default is FALSE.

Temp_Dir

Character string specifying the path for temporary storage of intermediate files.

Temp_Cleanup

logical, indicating whether to clean up temporary files. Defaults to TRUE.

Verbose

Logical. If TRUE, detailed output is printed. Default is FALSE.

Value

Depending on the value of ReturnData, either returns response curve data or NULL invisibly.

See also

Author

Ahmed El-Gabbas