Calculate the length of rivers in each Strahler order

This function processes EU-Hydro River Network Database to calculate the length of rivers in each Strahler number. The Strahler number is used as an index for river network classification, with higher numbers representing larger, more significant river segments. The function reads and processes zip-compressed geographic data (GPKG files), extracts relevant information about river segments, computes the length of rivers for each Strahler order per grid cell, and outputs the results both as raster files and RData objects. The calculated length represents the total length of rivers in each Strahler number or larger (e.g., for STRAHLER_5, the length of rivers with STRAHLER values of 5 or higher).

Usage

River_Length(
  FromHPC = TRUE,
  EnvFile = ".env",
  Filename = "EU_hydro_gpkg_eu.zip",
  Cleanup = FALSE
)

Arguments

FromHPC

Logical. Whether the processing is being done on an High-Performance Computing (HPC) environment, to adjust file paths accordingly. Default: TRUE.

EnvFile

Character. Path to the environment file containing paths to data sources. Defaults to .env.

Filename

Character. The name of the ZIP file containing the river network data. Default is "EU_hydro_gpkg_eu.zip".

Cleanup

Logical indicating whether to clean up temporary files from the Interim directory after finishing calculations. Default: FALSE.

Value

NULL. The function outputs processed files to the specified directories.

Details

The data provides at pan-European level a photo-interpreted river network, consistent of surface interpretation of water bodies (lakes and wide rivers), and a drainage model (also called Drainage Network), derived from EU-DEM, with catchments and drainage lines and nodes.

Data source: EU-Hydro River Network Database v013 | DOI | Download link

Temporal extent: 2006-2012; Format: Vector (GPKG); Size: 4 GB

Author

Ahmed El-Gabbas