Set up or stop parallel processing plan

Configures parallel processing with future::plan() or stops an existing plan. When stopping, it resets to sequential mode.

Usage

Set_parallel(
  NCores = 1L,
  Stop = FALSE,
  Cat = TRUE,
  Level = 0L,
  Future_maxSize = 8
)

Arguments

NCores

Integer. Number of cores to use. If NULL, defaults to sequential mode.

Stop

Logical. If TRUE, stops any parallel cluster and resets to sequential mode. If FALSE (default), sets up a new plan.

Cat

Logical. If TRUE (default), logs messages via CatTime().

Level

Integer. The logging level for CatTime. Default is 0.

Set_parallel

Numeric. Maximum allowed total size (in megabytes) of global variables identified. See “ argument of future::future.options for more details

Value

Invisible NULL. On Windows with NCores > 1, the cluster object (c1) is assigned to the parent environment and cleaned up automatically on exit.

Author

Ahmed El-Gabbas

Examples

if (FALSE) { # \dontrun{
  # Prepare working on parallel
  IASDT.R::Set_parallel(NCores = min(NCores, nrow(Beta_DF)), Level = 3)

  if (.Platform$OS.type == "windows") {
    on.exit(try(snow::stopCluster("c1"), silent = TRUE), add = TRUE)
  }
  on.exit(future::plan("future::sequential", gc = TRUE), add = TRUE)

  # Stopping cluster
  IASDT.R::Set_parallel(Stop = TRUE, Cat = TRUE, Level = 3)
} # }