Predict and evaluate cross-validated Hmsc models

This function computes predicted values of cross-validated Hmsc models in the testing cross-validation folders and evaluates the model performance using different metrics. In contrast to predict_maps, this function predicts only under current climate conditions and does not use clamping.

Usage

predict_maps_CV(
  model_dir = NULL,
  CV_name = NULL,
  CV_fold = NULL,
  n_cores = 8L,
  use_TF = TRUE,
  TF_environ = NULL,
  TF_use_single = FALSE,
  LF_n_cores = n_cores,
  LF_check = FALSE,
  LF_temp_cleanup = TRUE,
  LF_only = FALSE,
  LF_commands_only = FALSE,
  temp_cleanup = TRUE
)

Arguments

model_dir

Character. Path to the root directory where the cross-validated models were fitted.

CV_name

Character. Name of the cross-validation strategy used. Valid values are CV_Dist, CV_Large, or CV_SAC.

CV_fold

Integer. The cross-validation fold number.

n_cores

Integer. Number of CPU cores to use for parallel processing. Default: 8.

use_TF

Logical. Whether to use TensorFlow for calculations. Defaults to TRUE.

TF_environ

Character. Path to the Python environment. This argument is required if use_TF is TRUE under Windows. Defaults to NULL.

TF_use_single

Logical. Whether to use single precision for the TensorFlow calculations. Defaults to FALSE.

LF_n_cores

Integer. Number of cores to use for parallel processing of latent factor prediction. Defaults to 8L.

LF_check

Logical. If TRUE, the function checks if the output files are already created and valid. If FALSE, the function will only check if the files exist without checking their integrity. Default is FALSE.

LF_temp_cleanup

Logical. Whether to delete temporary files in the temp_dir directory after finishing the LF predictions.

LF_only

Logical. Whether to return the latent factor predictions only. Defaults to FALSE. This helps in predicting to new sites, allowing to predicting the latent factors only once, then the output can be loaded in other predictions when needed.

LF_commands_only

Logical. If TRUE, returns the command to run the Python script. Default is FALSE.

temp_cleanup

Logical. Whether to clean up temporary files. Defaults to TRUE.

Value

A tibble containing the prediction summary and file paths for output *.tif files.

See also

Author

Ahmed El-Gabbas