Merge model chains

This function merges model chains, checks for missing or incomplete model runs, and optionally prints information about incomplete models. It saves the processed model objects to disk.

Usage

mod_merge_chains(
  model_dir = NULL,
  n_cores = NULL,
  model_info_name = NULL,
  print_incomplete = TRUE,
  from_JSON = FALSE,
  out_extension = "qs2"
)

Arguments

model_dir

Character. Path to the root directory of the fitted model. Two folders will be created Model_Fitted and Model_Coda to store merged model and coda objects, respectively.

n_cores

Integer. Number of CPU cores to use for parallel processing.

model_info_name

Character. Name of the file (without extension) where the processed model information will be saved. If NULL, it overwrites the Model_Info.RData file in the model_dir directory. If provided, a new .RData file will be created with this name.

print_incomplete

Logical. Whether to print the names of models that were not successfully fitted to the console. Defaults to TRUE.

from_JSON

Logical. Whether to convert loaded models from JSON format before reading. Defaults to FALSE.

out_extension

Character. File extension (without dot) for the files containing the fitted model and coda objects. It can be either qs2 or RData. Defaults to qs2.

Value

The function does not return anything but saves the processed model information to disk.

Author

Ahmed El-Gabbas